requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
firebase-admin==6.4.0
python-dotenv==1.0.1
pydantic==2.6.0
